+++ src/main/java/roomescape/dao/JdbcReservationTimeDao.java
@@ -1,0 +1,62 @@
+package roomescape.dao;
+
+import java.util.List;
+import javax.sql.DataSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequest;
+
+@Repository
+public class JdbcReservationTimeDao implements ReservationTimeDao{
+
+    private final JdbcTemplate jdbcTemplate;
+    private final RowMapper<ReservationTime> reservationTimeMapper = (resultSet, row) ->
+            new ReservationTime(
+                    resultSet.getLong("id"),
+                    resultSet.getTime("start_at").toLocalTime()
+            );
+    private SimpleJdbcInsert reservationInsert;
+
+    public JdbcReservationTimeDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Autowired
+    public void setDataSource(DataSource dataSource) {
+        this.reservationInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("reservation_time")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    @Override
+    public Long insert(ReservationTimeRequest reservationTimeRequest) {
+        SqlParameterSource parameterSource = new BeanPropertySqlParameterSource(reservationTimeRequest);
+        Number id = reservationInsert.executeAndReturnKey(parameterSource);
+        return id.longValue();
+    }
+
+    @Override
+    public List<ReservationTime> findAllReservationTimes() {
+        String sql = "SELECT * FROM reservation_time";
+        return jdbcTemplate.query(sql, reservationTimeMapper);
+    }
+
+    @Override
+    public ReservationTime findById(Long id) {
+        String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, reservationTimeMapper, id);
+    }
+
+    @Override
+    public void delete(Long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+
+        jdbcTemplate.update(sql, id);
+    }
+}
